feat: Add export of key data
